// src/app/core/Models/classes/User/LoggedUserDto.ts
export class LoggedUserDto {
  userId: number = 0;
  firstName: string = '';
  lastName: string = '';
  userName: string = '';
  email: string = '';
  phoneNumber: number = 0;
  userTypeName: string = '';
  specialisationName: string = '';
  profileImage: string = '';
  qualification: string = '';
  registrationNumber: string = '';
  visitingCharge: number = 0;
}